requirements.txt
adal==1.2.2
ansiwrap==0.8.4
applicationinsights==0.11.9
appnope==0.1.0
attrs==19.3.0
azure-common==1.1.23
azure-graphrbac==0.61.1
azure-mgmt-authorization==0.60.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.0.0
azure-mgmt-resource==5.1.0
azure-mgmt-storage==6.0.0
azureml-automl-core==1.0.72.1
azureml-contrib-explain-model==1.0.65
azureml-contrib-notebook==1.0.72
azureml-core==1.0.72
azureml-dataprep==1.1.29
azureml-dataprep-native==13.1.0
azureml-defaults==1.0.72
azureml-explain-model==1.0.65
azureml-interpret==1.0.72.1
azureml-model-management-sdk==1.0.1b6.post1
azureml-pipeline==1.0.72
azureml-pipeline-core==1.0.72
azureml-pipeline-steps==1.0.72
azureml-sdk==1.0.72
azureml-telemetry==1.0.72
azureml-tensorboard==1.0.72
azureml-train==1.0.72
azureml-train-automl==1.0.72
azureml-train-core==1.0.72
azureml-train-restclients-hyperdrive==1.0.72
azureml-widgets==1.0.72
backcall==0.1.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bleach==3.1.0
boto==2.49.0
boto3==1.10.10
botocore==1.13.10
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
configparser==3.7.4
contextlib2==0.6.0.post1
cryptography==2.8
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
defusedxml==0.6.0
dill==0.3.1.1
distro==1.4.0
docker==4.1.0
docutils==0.15.2
dotnetcore2==2.1.10
entrypoints==0.3
fire==0.2.1
Flask==1.0.3
fusepy==3.0.1
future==0.18.2
gensim==3.8.1
gunicorn==19.9.0
idna==2.8
imageio==2.6.1
importlib-metadata==0.23
interpret-community==0.1.0.3.2
interpret-core==0.1.18
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.15.1
jeepney==0.4.1
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
json-logging-py==0.2
JsonForm==0.0.2
jsonpickle==1.2
jsonschema==3.1.1
JsonSir==0.0.2
jupyter-client==5.3.4
jupyter-core==4.6.1
keras2onnx==1.6.0
kiwisolver==1.1.0
liac-arff==2.4.0
lightgbm==2.3.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==7.2.0
msrest==0.6.10
msrestazure==0.6.2
nbconvert==5.6.1
nbformat==4.4.0
ndg-httpsclient==0.5.1
networkx==2.4
nimbusml==1.5.0
notebook==6.0.2
numpy==1.16.2
oauthlib==3.1.0
onnx==1.6.0
onnxconverter-common==1.6.0
onnxmltools==1.4.1
packaging==19.2
pandas==0.23.4
pandocfilters==1.4.2
papermill==1.2.1
parso==0.5.1
pathspec==0.6.0
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pmdarima==1.1.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
psutil==5.6.4
ptyprocess==0.6.0
pyarrow==0.11.1
pyasn1==0.4.7
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
pyOpenSSL==19.0.0
pyparsing==2.4.4
pyrsistent==0.15.5
python-dateutil==2.8.1
Python-EasyConfig==0.1.7
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
pyzmq==18.1.0
requests==2.22.0
requests-oauthlib==1.2.0
Resource==0.2.1
ruamel.yaml==0.15.89
s3transfer==0.2.1
scikit-image==0.16.2
scikit-learn==0.20.3
scipy==1.1.0
SecretStorage==3.1.1
Send2Trash==1.5.0
shap==0.29.3
six==1.13.0
skl2onnx==1.4.9
sklearn-pandas==1.7.0
smart-open==1.9.0
statsmodels==0.10.1
tenacity==5.1.5
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.4
textwrap3==0.9.2
tornado==6.0.3
tqdm==4.37.0
traitlets==4.3.3
typing-extensions==3.7.4.1
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.16.0
widgetsnbextension==3.5.1
zipp==0.6.0
